FIX: Collision reliability -- still needs some work but much improved.

+++ coresdk/src/coresdk/line_geometry.cpp
@@ -33,12 +33,12 @@ float line_length_squared(const line &l)
 
 line line_from(const vector_2d &v)
 {
-    return { {0,0}, {v.x, v.y} };
+    return { {0,0}, {static_cast<float>(v.x), static_cast<float>(v.y)} };
 }
 
 line line_from(const point_2d pt, const vector_2d &v)
 {
-    return { {pt.x, pt.y}, {pt.x + v.x, pt.y + v.y} };
+    return { {pt.x, pt.y}, {static_cast<float>(pt.x + v.x), static_cast<float>(pt.y + v.y)} };
 }
 
 bool line_intersection_point(const line &line1, const line &line2, point_2d &pt)
